+++ src/crypto/aes_ctr.cpp
@@ -113,10 +113,10 @@ AesCtr::Key AesCtr::generate_key() {
     // Use high-resolution clock for better randomness
     auto seed = std::chrono::high_resolution_clock::now().time_since_epoch().count();
     std::mt19937_64 gen(seed);
-    std::uniform_int_distribution<uint8_t> dist(0, 255);
+    std::uniform_int_distribution<uint16_t> dist(0, 255);
     
     for (auto& byte : key) {
-        byte = dist(gen);
+        byte = static_cast<uint8_t>(dist(gen));
     }
     
     return key;
@@ -128,10 +128,10 @@ AesCtr::IV AesCtr::generate_iv() {
     // Use high-resolution clock for better randomness
     auto seed = std::chrono::high_resolution_clock::now().time_since_epoch().count();
     std::mt19937_64 gen(seed);
-    std::uniform_int_distribution<uint8_t> dist(0, 255);
+    std::uniform_int_distribution<uint16_t> dist(0, 255);
     
     for (auto& byte : iv) {
-        byte = dist(gen);
+        byte = static_cast<uint8_t>(dist(gen));
     }
     
     return iv;
